Skeleton spiny ops.
