+++ nmv/skeleton/ops/__init__.py
@@ -31,3 +31,4 @@ from .skeleton_generic_ops import *
 from .skeleton_style_ops import *
 from .skeleton_verification_ops import *
 from .skeleton_soma_ops import *
+from .skeleton_spiny_ops import *
